+++ psutil/arch/windows/ntextapi.h
@@ -7,6 +7,17 @@
 #define __NTEXTAPI_H__
 #include <winternl.h>
 
+
+typedef struct {
+    LARGE_INTEGER IdleTime;
+    LARGE_INTEGER KernelTime;
+    LARGE_INTEGER UserTime;
+    LARGE_INTEGER DpcTime;
+    LARGE_INTEGER InterruptTime;
+    ULONG InterruptCount;
+} _SYSTEM_PROCESSOR_PERFORMANCE_INFORMATION;
+
+
 typedef enum _KTHREAD_STATE {
     Initialized,
     Ready,
@@ -19,6 +30,7 @@ typedef enum _KTHREAD_STATE {
     GateWait,
     MaximumThreadState
 } KTHREAD_STATE, *PKTHREAD_STATE;
+
 
 typedef enum _KWAIT_REASON {
     Executive = 0,
@@ -61,10 +73,12 @@ typedef enum _KWAIT_REASON {
     MaximumWaitReason = 37
 } KWAIT_REASON, *PKWAIT_REASON;
 
+
 typedef struct _CLIENT_ID {
     HANDLE UniqueProcess;
     HANDLE UniqueThread;
 } CLIENT_ID, *PCLIENT_ID;
+
 
 typedef struct _SYSTEM_THREAD_INFORMATION {
     LARGE_INTEGER KernelTime;
@@ -80,7 +94,9 @@ typedef struct _SYSTEM_THREAD_INFORMATION {
     KWAIT_REASON WaitReason;
 } SYSTEM_THREAD_INFORMATION, *PSYSTEM_THREAD_INFORMATION;
 
+
 typedef struct _TEB *PTEB;
+
 
 // private
 typedef struct _SYSTEM_EXTENDED_THREAD_INFORMATION {
@@ -93,6 +109,7 @@ typedef struct _SYSTEM_EXTENDED_THREAD_INFORMATION {
     ULONG_PTR Reserved3;
     ULONG_PTR Reserved4;
 } SYSTEM_EXTENDED_THREAD_INFORMATION, *PSYSTEM_EXTENDED_THREAD_INFORMATION;
+
 
 typedef struct _SYSTEM_PROCESS_INFORMATION2 {
     ULONG NextEntryOffset;
@@ -151,6 +168,7 @@ typedef struct _WINSTATION_INFO {
     FILETIME CurrentTime;
 } WINSTATION_INFO, *PWINSTATION_INFO;
 
+
 typedef BOOLEAN (WINAPI * PWINSTATIONQUERYINFORMATIONW)
                  (HANDLE,ULONG,WINSTATIONINFOCLASS,PVOID,ULONG,PULONG);
 
@@ -162,6 +180,7 @@ typedef BOOLEAN (WINAPI * PWINSTATIONQUERYINFORMATIONW)
  */
 typedef LONG NTSTATUS;
 
+
 typedef NTSTATUS (NTAPI *_NtQueryInformationProcess)(
     HANDLE ProcessHandle,
     DWORD ProcessInformationClass,
@@ -170,12 +189,14 @@ typedef NTSTATUS (NTAPI *_NtQueryInformationProcess)(
     PDWORD ReturnLength
 );
 
+
 typedef NTSTATUS (NTAPI *_NtSetInformationProcess)(
     HANDLE ProcessHandle,
     DWORD ProcessInformationClass,
     PVOID ProcessInformation,
     DWORD ProcessInformationLength
 );
+
 
 typedef enum _PROCESSINFOCLASS2 {
     _ProcessBasicInformation,
@@ -223,6 +244,7 @@ typedef enum _PROCESSINFOCLASS2 {
     ProcessImageInformation,
     MaxProcessInfoClass
 } PROCESSINFOCLASS2;
+
 
 #define PROCESSINFOCLASS PROCESSINFOCLASS2
 #define ProcessBasicInformation _ProcessBasicInformation